Use separate scheduler for welcome emails
[MAILPOET-2321]
