+++ lib/DI/ContainerConfigurator.php
@@ -182,6 +182,7 @@ class ContainerConfigurator implements IContainerConfigurator {
     // Newsletter
     $container->autowire(\MailPoet\Newsletter\AutomatedLatestContent::class)->setPublic(true);
     $container->autowire(\MailPoet\Newsletter\NewslettersRepository::class);
+    $container->autowire(\MailPoet\Newsletter\Scheduler\WelcomeScheduler::class);
     // Util
     $container->autowire(\MailPoet\Util\Cookies::class);
     $container->autowire(\MailPoet\Util\Url::class)->setPublic(true);
